+++ packages/zfast/src/index.ts
@@ -1,3 +1,4 @@
 export { render } from "./client/render";
 export { defineConfig } from "./defineConfig";
 export { ICodeItem, IEntryImport } from "./types";
+export { App } from "./app";